+++ assets/api/user/profileAxiosApi.ts
@@ -4,9 +4,9 @@ import { UserProfile, UpdateUserProfile } from './userTypes'
 import { NextApiRequest } from 'next'
 import { createAxiosServerInstance, instance } from '../instance'
 
-const getProfileFromServer = async (req: NextApiRequest) => {
+const getProfileFromServer = async (req: NextApiRequest, id: number) => {
   const axiosInstance = createAxiosServerInstance(req)
-  const response = await axiosInstance.get<AxiosResponse<UserProfile>>(userRouts.profile)
+  const response = await axiosInstance.get<AxiosResponse<UserProfile>>(userRouts.profile + id)
   return response.data
 }
