+++ pages/profile/index.tsx
@@ -21,9 +21,11 @@ import { withAuth } from '@/utils/getServerSideProps/withAuth'
 import { useEffect, useState } from 'react'
 import { ProfileType } from '@/components/Profile-Settings/profileSettingsTypes'
 import { useScrollEffect } from '../../hooks/useScrollEffect'
+import MyProfile from '@/components/MyProfile/MyProfile'
 
 export const getServerSideProps: GetServerSideProps = withAuth(async ({ req }) => {
-  const userProfile = await axiosAPI.profile.getProfileFromServer(req as NextApiRequest)
+  const isAuth = await axiosAPI.auth.meServer(req as NextApiRequest)
+  const userProfile = await axiosAPI.profile.getProfileFromServer(req as NextApiRequest, isAuth.userId)
 
   if (!userProfile) {
     return {
@@ -38,59 +40,13 @@ export const getServerSideProps: GetServerSideProps = withAuth(async ({ req }) =
   }
 })
 
-const MyProfile = ({ userProfile }: ProfileType) => {
-  console.log(userProfile)
-
-  const [images, setImagesContent] = useState(['fff', 'hhhh'])
-
-  const router = useRouter()
-
-  useScrollEffect(setImagesContent, images)
-
-  const handleClick = () => {
-    router.push(userRouts.profileSettings)
-  }
-
+const MyProfilePage = ({ userProfile }: ProfileType) => {
   return (
     <PageWrapper>
-      <div className={s.main}>
-        <div className={s.header}>
-          <Image className={s.avatar} width={204} height={204} src={userProfile.avatars[0].url} alt="Avatar" />
-          <div className={s.info}>
-            <div className={s.blockTop}>
-              <div className={s.name}>{userProfile.firstName}</div>
-              <div className={s.settings}>
-                <Button color={BUTTON_COLORS.BASIC} onClick={handleClick} text="Profile Settings"></Button>
-              </div>
-            </div>
-            <div className={s.centre}>
-              <div className={s.item}>
-                <span>2 218</span>
-                <span>Following</span>
-              </div>
-              <div className={s.item}>
-                <span>2 358</span>
-                <span>Followers</span>
-              </div>
-              <div className={s.item}>
-                <span>2 764</span>
-                <span>Publications</span>
-              </div>
-            </div>
-            <div className={s.blockBottom}>
-              <p className={s.text}>{userProfile.aboutMe}</p>
-            </div>
-          </div>
-        </div>
-        <div className={s.post}>
-          {
-            // тут посты отображаются
-          }
-        </div>
-      </div>
+      <MyProfile userProfile={userProfile} />
     </PageWrapper>
   )
 }
 
-MyProfile.getLayout = getSideBarLayout
-export default MyProfile
+MyProfilePage.getLayout = getSideBarLayout
+export default MyProfilePage
